Complete main_patterns_04 homework

+++ src/VO/Vector.php
@@ -5,7 +5,7 @@ namespace Mettlive\SquareSolver\VO;
 class Vector
 {
 
-    public function __construct(public int $x, public int $y)
+    public function __construct(public float $x, public float $y)
     {
     }
 }
